# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StellarRoute API
  version: 0.1.0
  description: |
    REST API for DEX aggregation on the Stellar Network.

    StellarRoute aggregates liquidity from Stellar's SDEX (Stellar Distributed
    Exchange) and Soroban AMM pools to find the best available price for any
    token pair.

    ## Asset identifiers

    Assets are encoded in path segments using one of three formats:

    | Format | Example | Meaning |
    |--------|---------|---------|
    | `native` | `native` | XLM (Stellar's native asset) |
    | `CODE` | `USDC` | Issued asset — code only (omits issuer) |
    | `CODE:ISSUER` | `USDC:GA5ZSEJ...` | Issued asset — fully qualified |

    ## Error handling

    All errors return a JSON body with `error` (machine-readable code) and
    `message` (human-readable description).  See the `ErrorResponse` schema
    and the **Error codes** section below for the full list.

    ## Rate limiting

    Endpoints are rate-limited per IP address.  When the limit is exceeded the
    API returns **429 Too Many Requests** with `X-RateLimit-*` headers.

    ## Request correlation

    Clients may optionally send `X-Request-ID` with any request. The API will
    reuse that value, or generate a UUID when the header is omitted, and echo
    the final request ID back in the `X-Request-ID` response header.

    ## Versioning and deprecation

    `/api/v1/*` is in deprecation mode. Responses from v1 routes include
...
    notice period and migration guidance.
...
servers:
...
paths:
...
  /api/v1/price-history/{base}/{quote}:
    get:
      operationId: get_price_history
      summary: Get 24h price history
      description: |
        Returns a lightweight 24h sparkline series for the selected trading pair.

        The backend derives the series from `orderbook_snapshots.mid_price`
        using hourly aggregation buckets. If no snapshots are available in the
        trailing 24h window, the endpoint returns an empty `points` array.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
      responses:
        "200":
          description: 24h price history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistoryResponse"
              example:
                base_asset:
                  asset_type: native
                  asset_code: null
                  asset_issuer: null
                quote_asset:
                  asset_type: credit_alphanum4
                  asset_code: USDC
                  asset_issuer: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
                window: "24h"
                source: orderbook_snapshots.mid_price
                generated_at: 1740312000000
                points:
                  - timestamp: 1740225600000
                    price: "0.1051000"
                  - timestamp: 1740229200000
                    price: "0.1054000"
        "400":
          description: Invalid asset identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trading pair not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    BaseAsset:
      name: base
      in: path
      required: true
      description: |
        Base asset identifier. One of:
        - `native` — XLM
        - `CODE` — issued asset code only (e.g. `USDC`)
        - `CODE:ISSUER` — fully-qualified issued asset (e.g. `USDC:GA5Z...`)
      schema:
        type: string
        example: native
    QuoteAsset:
      name: quote
      in: path
      required: true
      description: |
        Quote asset identifier. One of:
        - `native` — XLM
        - `CODE` — issued asset code only (e.g. `USDC`)
        - `CODE:ISSUER` — fully-qualified issued asset (e.g. `USDC:GA5Z...`)
      schema:
        type: string
        example: USDC
...
  schemas:
...
    AssetInfo:
      type: object
      required: [asset_type]
      description: Stellar asset descriptor
      properties:
        asset_type:
          type: string
          description: Stellar asset type
          enum: [native, credit_alphanum4, credit_alphanum12]
          example: credit_alphanum4
        asset_code:
          type: string
          nullable: true
          description: Asset code (null for native XLM)
          example: USDC
        asset_issuer:
          type: string
          nullable: true
          description: G-address of the issuing account (null for native XLM)
          example: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
...
    PriceHistoryPoint:
      type: object
      required: [timestamp, price]
      description: Aggregated mid-price sample for a 24h sparkline
      properties:
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds for the sampling bucket
          example: 1740225600000
        price:
          type: string
          description: Average mid-price for the bucket
          example: "0.1051000"
    PriceHistoryResponse:
      type: object
      required: [base_asset, quote_asset, window, source, generated_at, points]
      description: Historical price series derived from orderbook snapshots
      properties:
        base_asset:
          $ref: "#/components/schemas/AssetInfo"
        quote_asset:
          $ref: "#/components/schemas/AssetInfo"
        window:
          type: string
          description: Fixed time window for the series
          example: "24h"
        source:
          type: string
          description: Backend data source contract
          example: orderbook_snapshots.mid_price
        generated_at:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds when the series was generated
          example: 1740312000000
        points:
          type: array
          items:
            $ref: "#/components/schemas/PriceHistoryPoint"
...
    ErrorResponse:
      type: object
      required: [error, message]
      description: Standard API error format
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: validation_error
        message:
          type: string
          description: Human-readable error description
          example: "Amount must be greater than zero"
        details:
          type: object
          nullable: true
          description: Optional structured context about the error
          example:
            parameter: slippage_bps
            min: 0
            max: 10000

  # Authentication is planned for a future release.
  # securitySchemes:
  #   bearerAuth:
  #     type: http
  #     scheme: bearer
  #     bearerFormat: JWT
  #     description: |
  #       API key authentication (Phase 2).
  #       Pass a JWT obtained from the /auth/token endpoint.